# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customer/me:
    get:
      operationId: getCustomerProfile
      tags: [customer]
      summary: The customer this browser is signed in as, if any
      responses:
        "200":
          description: Profile, or authenticated false with empty fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
    patch:
      operationId: updateCustomerProfile
      tags: [customer]
      summary: Save name and phone, creating the customer on first save
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerProfileInput"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "400":
          description: Invalid phone, or a verified phone cannot be changed here
  /customer/me/addresses:
    put:
      operationId: replaceCustomerAddresses
      tags: [customer]
      summary: Replace the saved delivery addresses, most recent first
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerAddressesInput"
      responses:
        "200":
          description: Saved profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
  /customer/login:
    get:
      operationId: getCustomerLoginLink
      tags: [customer]
      summary: The Telegram link that starts phone verification
      responses:
        "200":
          description: Open it, share the phone number, and the bot sends a code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerLoginLink"
        "503":
          description: The Telegram bot is not configured
    post:
      operationId: verifyCustomerLogin
      tags: [customer]
      summary: Verify the phone with the code the bot sent, and sign in
      description: >-
        If the verified phone already belongs to a customer (another device),
        this browser signs in as that customer and brings its orders,
        addresses and chat along.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerLoginInput"
      responses:
        "200":
          description: Verified and signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerProfile"
        "401":
          description: Wrong or expired code
        "429":
          description: Too many attempts
...
components:
  schemas:
...
    DeliveryAddress:
      type: object
      required: [dom, xonadon]
      properties:
        dom:
          type: string
          maxLength: 10
        xonadon:
          type: string
          maxLength: 10
    CustomerProfile:
      type: object
      required: [authenticated, name, phone, phone_verified, addresses]
      properties:
        authenticated:
          type: boolean
        name:
          type: string
        phone:
          type: string
        phone_verified:
          type: boolean
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/DeliveryAddress"
    CustomerProfileInput:
      type: object
      properties:
        name:
          type: string
          maxLength: 80
        phone:
          type: string
          maxLength: 32
    CustomerAddressesInput:
      type: object
      required: [addresses]
      properties:
        addresses:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/DeliveryAddress"
    CustomerLoginLink:
      type: object
      required: [url]
      properties:
        url:
          type: string
    CustomerLoginInput:
      type: object
      required: [phone, code]
      properties:
        phone:
          type: string
          minLength: 7
          maxLength: 32
        code:
          type: string
          pattern: "^[0-9]{6}$"
...