# lib/api-zod/src/generated/types/customerAddressesInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryAddress } from './deliveryAddress';

export interface CustomerAddressesInput {
  /** @maxItems 5 */
  addresses: DeliveryAddress[];
}