# lib/api-zod/src/generated/types/customerLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerLoginInput {
  /**
     * @minLength 7
     * @maxLength 32
     */
  phone: string;
  /** @pattern ^[0-9]{6}$ */
  code: string;
}