# lib/api-zod/src/generated/types/customerProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeliveryAddress } from './deliveryAddress';

export interface CustomerProfile {
  authenticated: boolean;
  name: string;
  phone: string;
  phone_verified: boolean;
  addresses: DeliveryAddress[];
}