# lib/api-zod/src/generated/types/customerProfileInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerProfileInput {
  /** @maxLength 80 */
  name?: string;
  /** @maxLength 32 */
  phone?: string;
}