# lib/api-zod/src/generated/types/deliveryAddress.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DeliveryAddress {
  /** @maxLength 10 */
  dom: string;
  /** @maxLength 10 */
  xonadon: string;
}